Put download buttons inside iframes
Some browsers (Chromium for sure, but probably others too) don't behave
well when combining EventSource and file downloads. When an user clicks
on a button to trigger a file donwload (i.e. a file served with a
"Content-Disposition: attachment" header) the browser closes the
EventSource connection (without even sending a 'close' event), navigates
to the new page, realizes that it's a file download, opens a prompt to
ask the user to open/save the file and (simultaneously) goes back to the
previous page without reopening the EventSource connection or sending
any kind of events.

Our application wants to always be connected to the server and to know
when it is not, to alert the user and, in case, ask him to reconnect.
Therefore this behavior is unacceptable.

The best fix I came up with is to put the download buttons inside
iframes, so that only the browsing context of the iframe is affected
when the user clicks on a button, and not the main context (where the
EventSource is bound to). Any suggestion for a better fix is welcome.